gpkg2: T_roads_railways

+++ gpkg2.sh
@@ -118,7 +118,7 @@ cd "$dir"
 #                                             #
 # =========================================== #
 
-# 19"
+# 12"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$m_gpk" "$fconf"`
@@ -164,7 +164,7 @@ fi
 #                                    #
 # ================================== #
 
-# 34"
+# 33"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$s_gpk" "$fconf"`
@@ -197,7 +197,7 @@ fi
 #                                                          #
 # ======================================================== #
 
-# 8'57"
+# 9'58"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$d_gpk" "$fconf"`
@@ -243,7 +243,7 @@ then
 	# Garapenera edo ekoizpenera kopiatu / Copiar a desarrollo o a producción
 	cp_gpk "$typ01" "$d_gpk"
 	msg " - ${typ01}"
-	rm "$f02" 2> /dev/null
+42m "$f02" 2> /dev/null
 fi
 
 # ==================================== #
@@ -252,7 +252,7 @@ fi
 #                                      #
 # ==================================== #
 
-# 9'04"
+# 9'42"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$e_gpk" "$fconf"`
@@ -298,7 +298,7 @@ then
 	# Garapenera edo ekoizpenera kopiatu / Copiar a desarrollo o a producción
 	cp_gpk "$typ01" "$e_gpk"
 	msg " - ${typ01}"
-	rm "$f02" 2> /dev/null
+5rm "$f02" 2> /dev/null
 fi
 
 # ================================================================================ #
@@ -307,7 +307,7 @@ fi
 #                                                                                  #
 # ================================================================================ #
 
-# 1'36"
+# 1'56"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$k_gpk" "$fconf"`
@@ -354,7 +354,7 @@ fi
 #                                                   #
 # ================================================= #
 
-# 9'42"
+# 12'39"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$v_gpk" "$fconf"`
@@ -401,7 +401,7 @@ fi
 #                             #
 # =========================== #
 
-# 17"
+# 36"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$c_gpk" "$fconf"`
@@ -448,7 +448,7 @@ fi
 #                                              #
 # ============================================ #
 
-# 51"
+# 43"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$i_gpk" "$fconf"`
@@ -495,7 +495,7 @@ fi
 #                                                                     #
 # =================================================================== #
 
-# 11"
+# 31"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$z_gpk" "$fconf"`
@@ -542,7 +542,7 @@ fi
 #                                                                      #
 # ==================================================================== #
 
-# 36"
+# 26"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$g_gpk" "$fconf"`
@@ -579,6 +579,53 @@ then
 
 	# Garapenera edo ekoizpenera kopiatu / Copiar a desarrollo o a producción
 	cp_gpk "$typ01" "$g_gpk"
+	msg " - ${typ01}"
+	rm "$f02" 2> /dev/null
+fi
+
+# ======================================================================== #
+#                                                                          #
+# 11. t_roads_railways (errepidea eta trenbidea / carretera y ferrocarril) #
+#                                                                          #
+# ======================================================================== #
+
+# ??"
+
+# Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
+vconf=`grep "$t_gpk" "$fconf"`
+IFS='|' read -a aconf <<< "$vconf"
+typ01="${aconf[0]}"
+gpk01="${aconf[1]}"
+des01="${t_des[0]} - ${t_des[1]} - ${t_des[2]}"
+if [ "$t_gpk" = "$gpk01" ] && ([ $typ01 = "1" ] || [ "$typ01" = "2" ])
+then
+	let i1=$i1+1
+	msg "${i1}/${nf}: $(date '+%Y-%m-%d %H:%M:%S') - $gpk01 - ${t_des[0]}\c"
+	f01="${tmpd}/${t_gpk}_01.gpkg"
+	c01="${tmpd}/${t_gpk}_01.csv"
+	f02="${tmpd}/${t_gpk}.gpkg"
+
+	# Oinarrizko datuak / Datos básicos
+	rm "$f01" 2> /dev/null
+	ogr2ogr -f "GPKG" -s_srs "EPSG:25830" -t_srs "EPSG:25830" "$f01" OCI:${con1}:${tpl} -nln "$t_gpk" -lco DESCRIPTION="$des01" -sql "$t_sql_01"
+
+	# more_info
+	rm "$c01" 2> /dev/null
+	sql_more_info2 "$c01" "$t_sql_02"
+	rm "$f02" 2> /dev/null
+	ogr2ogr -f "GPKG" -update -s_srs "EPSG:25830" -t_srs "EPSG:25830" -nln "${t_gpk}_more_info" -lco DESCRIPTION="${des01} more info" "$f01" "$c01"
+	rm "$c01" 2> /dev/null
+
+	# Behin betiko GPKGa / GPKG definitivo
+	rm "$f02" 2> /dev/null
+	ogr2ogr -f "GPKG" -s_srs "EPSG:25830" -t_srs "EPSG:25830" -nln "$t_gpk" -lco DESCRIPTION="$des01" -sql "$t_sql_03" "$f02" "$f01"
+	rm "$f01" 2> /dev/null
+
+	# Eremuak berrizendatu / Renombrar campos
+	rfl "$f02" "$t_gpk"
+
+	# Garapenera edo ekoizpenera kopiatu / Copiar a desarrollo o a producción
+	cp_gpk "$typ01" "$t_gpk"
 	msg " - ${typ01}"
 	rm "$f02" 2> /dev/null
 fi
